# src/main/java/com/complus/internalbooking/swagger/V1Api.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.36).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.complus.internalbooking.swagger;

import com.complus.internalbooking.swagger.modal.CreateTradeRequest;
import com.complus.internalbooking.swagger.modal.CreateTradeResponse;
import com.complus.internalbooking.swagger.modal.GenerateReportRequest;
import io.swagger.annotations.*;

import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import javax.validation.Valid;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2022-12-09T12:20:30.692Z[GMT]")
@Api(value = "internal-domain/v1", description = "the v1 API")
@Validated
public interface V1Api {

    @ApiOperation(value = "Post create trade", notes = "", response = CreateTradeResponse.class, tags={ "Internal Booking" })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = CreateTradeResponse.class) })
    @RequestMapping(value = "/v1/trade/create",
        produces = { "application/json" }, 
        consumes = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<CreateTradeResponse> postCreateTrade(@ApiParam(value = "Body for create trade") @Valid @RequestBody CreateTradeRequest body);


    @ApiOperation(value = "Post generate trade report", notes = "", tags={ "Internal Booking" })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "successful operation") })
    @RequestMapping(value = "/v1/trade/report",
        consumes = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<Void> postGenerateReport(@ApiParam(value = "Body for generate trade report", required=true) @Valid @RequestBody GenerateReportRequest body);

}